requirements.txt
torchcrepe==0.0.19
torchfcpe==0.0.4
bitsandbytes>=0.43.1
uvicorn>=0.20.0
scipy>=1.13.0
scikit-learn>=1.3.1
rich>=13.6.0
deepfilternet>=0.5.6
faiss-cpu==1.7.2
fairseq==0.12.2
numpy>=1.23.5
rich==13.6.0
requests==2.31.0
pyworld==0.3.2
mutagen==1.47.0
lomo-optim==0.1.1
ffmpeg-python==0.2.0
fastapi==0.108.0
came-pytorch==0.1.3
py-cpuinfo==9.0.0
